autonomous-robotic-arm-rl/requirements.txt
gymnasium==1.2.3
stable-baselines3==2.7.1
mujoco==3.4.0
torch>=2.3
numpy
pyyaml
tensorboard
opencv-python